+++ tests/test_base_device.py
@@ -508,6 +508,9 @@ class TestKwargsHandling:
         assert dev._kwargs["execution"]["init_qubits"] is False
 
     def test_no_error_is_raised_if_transpilation_options_are_passed(self):
+        """Tests that when transpilation options are passed in, they are properly
+        handled without error"""
+
         dev = QiskitDevice2(
             wires=2,
             backend=backend,
@@ -1147,13 +1150,25 @@ class TestExecution:
         def circuit(x):
             qml.RX(x, wires=[0])
             qml.CNOT(wires=[0, 1])
-            return qml.sample(), qml.sample(qml.Y(0))
+            return (
+                qml.sample(),
+                qml.sample(qml.Y(0)),
+                qml.expval(qml.X(1)),
+                qml.var(qml.Y(0)),
+                qml.counts(),
+            )
 
         @qml.qnode(qiskit_dev)
         def qiskit_circuit(x):
             qml.RX(x, wires=[0])
             qml.CNOT(wires=[0, 1])
-            return qml.sample(), qml.sample(qml.Y(0))
+            return (
+                qml.sample(),
+                qml.sample(qml.Y(0)),
+                qml.expval(qml.X(1)),
+                qml.var(qml.Y(0)),
+                qml.counts(),
+            )
 
         res = circuit(np.pi / 2)
         qiskit_res = qiskit_circuit(np.pi / 2)