+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 ### Improvements 🛠
 
+* Updated `load_qasm` to take the optional kwarg `measurements` which get performed at the end of the loaded circuit and `load_qasm` can now detect mid-circuit measurements from `qasm`.
+[(#555)](https://github.com/PennyLaneAI/pennylane-qiskit/pull/555)
+
+* Improvements have been made to load circuits with `SwitchCaseOp` gates with default case.
+  [(#514)](https://github.com/PennyLaneAI/pennylane-qiskit/pull/514)
+
 ### Breaking changes 💔
 * Support has been removed for Qiskit versions below 0.46. The minimum required version for Qiskit is now 1.0. 
   If you want to continue to use older versions of Qiskit with the plugin, please use version 0.36 of 
@@ -29,6 +35,9 @@
 
 This release contains contributions from (in alphabetical order):
 Austin Huang
+
+Utkarsh Azad
+Mashhood Khan
 
 ---
 # Release 0.36.0